move MitigationSegments component to core
