+++ src/analysis/retail/monk/brewmaster/modules/core/MajorDefensives/config.ts
@@ -10,4 +10,4 @@ export const MAJOR_ANALYZERS = [
   DampenHarm,
   DiffuseMagic,
   ZenMeditation,
-];
+] as const;
